moved catalog fetch method
